gemini-analyst/requirements_ga.txt
Flask
google-genai
requests # get news
python-dotenv